update base class, extend base class and use on registration flow
