+++ src/app/features/registration/find-workplace-address/find-workplace-address.component.ts
@@ -1,120 +1,34 @@
-import { HttpErrorResponse } from '@angular/common/http';
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { Component } from '@angular/core';
+import { FormBuilder } from '@angular/forms';
 import { Router } from '@angular/router';
-import { ErrorDefinition, ErrorDetails } from '@core/model/errorSummary.model';
 import { LocationSearchResponse } from '@core/model/location.model';
 import { BackService } from '@core/services/back.service';
 import { ErrorSummaryService } from '@core/services/error-summary.service';
-import { RegistrationService } from '@core/services/registration.service';
-import { Subscription } from 'rxjs';
 import { LocationService } from '@core/services/location.service';
+import { RegistrationService } from '@core/services/registration.service';
+import { FindWorkplaceAddress } from '@features/workplace-find-and-select/find-workplace-address/find-workplace-address';
 
 @Component({
   selector: 'app-find-workplace-address',
   templateUrl: './find-workplace-address.component.html',
 })
-export class FindWorkplaceAddressComponent implements OnInit, OnDestroy {
-  public form: FormGroup;
-  public serverError: string;
-  public submitted = false;
-  private formErrorsMap: Array<ErrorDetails>;
-  private serverErrorsMap: Array<ErrorDefinition>;
-  private subscriptions: Subscription = new Subscription();
-
+export class FindWorkplaceAddressComponent extends FindWorkplaceAddress {
   constructor(
-    private backService: BackService,
-    private errorSummaryService: ErrorSummaryService,
-    private fb: FormBuilder,
-    private locationService: LocationService,
     private registrationService: RegistrationService,
-    private router: Router
-  ) {}
-
-  ngOnInit() {
-    this.setupForm();
-    this.setupFormErrorsMap();
-    this.setupServerErrorsMap();
-    this.setBackLink();
+    protected backService: BackService,
+    protected errorSummaryService: ErrorSummaryService,
+    protected formBuilder: FormBuilder,
+    protected locationService: LocationService,
+    protected router: Router
+  ) {
+    super(backService, errorSummaryService, formBuilder, locationService, router);
   }
 
-  get getPostcode() {
-    return this.form.get('postcode');
+  protected init(): void {
+    this.flow = '/registration';
   }
 
-  private setupForm(): void {
-    this.form = this.fb.group({
-      postcode: ['', [Validators.required, Validators.maxLength(8)]],
-    });
-  }
-
-  private setupFormErrorsMap(): void {
-    this.formErrorsMap = [
-      {
-        item: 'postcode',
-        type: [
-          {
-            name: 'required',
-            message: 'Please enter a postcode.',
-          },
-          {
-            name: 'maxlength',
-            message: 'Invalid postcode.',
-          },
-        ],
-      },
-    ];
-  }
-
-  private setupServerErrorsMap(): void {
-    this.serverErrorsMap = [
-      {
-        name: 400,
-        message: 'No results found.',
-      },
-      {
-        name: 404,
-        message: 'Invalid postcode.',
-      }
-    ];
-  }
-
-  private getAddressesByPostCode(): void {
-    this.subscriptions.add(
-      this.locationService.getAddressesByPostCode(this.getPostcode.value).subscribe(
-        (data: LocationSearchResponse) => {
-          this.registrationService.locationAddresses$.next(data.postcodedata);
-          this.router.navigate(['/registration/select-workplace-address']);
-        },
-        (error: HttpErrorResponse) => {
-          this.serverError = this.errorSummaryService.getServerErrorMessage(error.status, this.serverErrorsMap);
-          this.errorSummaryService.scrollToErrorSummary();
-        }
-      )
-    );
-  }
-
-  public onSubmit(): void {
-    this.submitted = true;
-    this.errorSummaryService.syncFormErrorsEvent.next(true);
-
-    if (this.form.valid) {
-      this.getAddressesByPostCode();
-    } else {
-      this.errorSummaryService.scrollToErrorSummary();
-    }
-  }
-
-  private setBackLink(): void {
-    this.backService.setBackLink({ url: ['/registration/select-workplace-address'] });
-  }
-
-  public getFirstErrorMessage(item: string): string {
-    const errorType = Object.keys(this.form.get(item).errors)[0];
-    return this.errorSummaryService.getFormErrorMessage(item, errorType, this.formErrorsMap);
-  }
-
-  ngOnDestroy(): void {
-    this.subscriptions.unsubscribe();
+  protected onSuccess(data: LocationSearchResponse): void {
+    this.registrationService.locationAddresses$.next(data.postcodedata);
   }
 }
