+++ src/app/features/workplace-find-and-select/find-workplace-address/find-workplace-address.ts
@@ -1,7 +1,120 @@
-import { OnInit } from '@angular/core';
+import { HttpErrorResponse } from '@angular/common/http';
+import { OnDestroy, OnInit } from '@angular/core';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
+import { ErrorDefinition, ErrorDetails } from '@core/model/errorSummary.model';
+import { LocationSearchResponse } from '@core/model/location.model';
+import { BackService } from '@core/services/back.service';
+import { ErrorSummaryService } from '@core/services/error-summary.service';
+import { LocationService } from '@core/services/location.service';
+import { Subscription } from 'rxjs';
 
-export class FindWorkplaceAddress implements OnInit {
-  constructor() {}
+export class FindWorkplaceAddress implements OnInit, OnDestroy {
+  protected serverErrorsMap: Array<ErrorDefinition>;
+  protected subscriptions: Subscription = new Subscription();
+  protected flow: string;
+  public form: FormGroup;
+  public formErrorsMap: Array<ErrorDetails>;
+  public serverError: string;
+  public submitted = false;
 
-  ngOnInit() {}
+  constructor(
+    protected backService: BackService,
+    protected errorSummaryService: ErrorSummaryService,
+    protected formBuilder: FormBuilder,
+    protected locationService: LocationService,
+    protected router: Router
+  ) {}
+
+  ngOnInit() {
+    this.setupForm();
+    this.setupFormErrorsMap();
+    this.setupServerErrorsMap();
+    this.init();
+    this.setBackLink();
+  }
+
+  get getPostcode() {
+    return this.form.get('postcode');
+  }
+
+  protected init(): void {}
+
+  private setupForm(): void {
+    this.form = this.formBuilder.group({
+      postcode: ['', [Validators.required, Validators.maxLength(8)]],
+    });
+  }
+
+  protected setupFormErrorsMap(): void {
+    this.formErrorsMap = [
+      {
+        item: 'postcode',
+        type: [
+          {
+            name: 'required',
+            message: 'Please enter a postcode.',
+          },
+          {
+            name: 'maxlength',
+            message: 'Invalid postcode.',
+          },
+        ],
+      },
+    ];
+  }
+
+  protected setupServerErrorsMap(): void {
+    this.serverErrorsMap = [
+      {
+        name: 400,
+        message: 'No results found.',
+      },
+      {
+        name: 404,
+        message: 'Invalid postcode.',
+      },
+    ];
+  }
+
+  protected getAddressesByPostCode(): void {
+    this.subscriptions.add(
+      this.locationService.getAddressesByPostCode(this.getPostcode.value).subscribe(
+        (data: LocationSearchResponse) => {
+          this.onSuccess(data);
+          this.router.navigate([`${this.flow}/select-workplace-address`]);
+        },
+        (error: HttpErrorResponse) => {
+          this.serverError = this.errorSummaryService.getServerErrorMessage(error.status, this.serverErrorsMap);
+          this.errorSummaryService.scrollToErrorSummary();
+        }
+      )
+    );
+  }
+
+  protected onSuccess(data: LocationSearchResponse): void {}
+
+  public onSubmit(): void {
+    this.submitted = true;
+    this.errorSummaryService.syncFormErrorsEvent.next(true);
+
+    if (this.form.valid) {
+      this.getAddressesByPostCode();
+    } else {
+      this.errorSummaryService.scrollToErrorSummary();
+    }
+  }
+
+  protected setBackLink(): void {
+    this.backService.setBackLink({ url: [`${this.flow}/select-workplace-address`] });
+  }
+
+  public getFirstErrorMessage(item: string): string {
+    const errorType = Object.keys(this.form.get(item).errors)[0];
+    return this.errorSummaryService.getFormErrorMessage(item, errorType, this.formErrorsMap);
+  }
+
+  ngOnDestroy(): void {
+    this.subscriptions.unsubscribe();
+  }
 }
